server/requirements.txt
openenv-core>=0.1.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
fastmcp>=0.1.0
